agents/agentic-rag/rag-types/prebuilt-agentic-rag-poc/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
langgraph